move course sync to a dedicated file
